CMN-293: remove stackability on digital list
