+++ src/apps/loan-list/list/loan-list-items.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState } from "react";
+import React, { FC } from "react";
 import { removeLoansWithDuplicateDueDate } from "../utils/helpers";
 import StackableMaterial from "../materials/stackable-material/stackable-material";
 import { ListView } from "../../../core/utils/types/list-view";
@@ -10,7 +10,7 @@ import { useText } from "../../../core/utils/text";
 interface LoanListItemProps {
   loans: LoanType[];
   view: ListView;
-  dueDates: string[];
+  dueDates?: string[];
   pageSize: number;
 }
 
@@ -21,16 +21,7 @@ const LoanListItems: FC<LoanListItemProps> = ({
   pageSize
 }) => {
   const t = useText();
-  const [localDueDates, setLocalDueDates] = useState<Array<string | null>>([]);
-
-  useEffect(() => {
-    if (view === "stacked") {
-      // Publizon material sometimes have due dates that are "null"
-      // These should also be displayed stacked
-      setLocalDueDates([...dueDates, null]);
-    }
-  }, [dueDates, view]);
-
+  console.log(loans);
   return (
     // explanation for screen readers used in additional-materials-button
     // It is located here to avoid duplicate ids in the dom
@@ -41,10 +32,11 @@ const LoanListItems: FC<LoanListItemProps> = ({
       >
         {t("loanListDueDateModalAriaLabelText")}
       </div>
-      {view === "stacked" &&
-        localDueDates.map((uniqueDueDate: string | null) => {
+      {view === "stack" &&
+        dueDates &&
+        dueDates.map((uniqueDueDate: string) => {
           // Stack items:
-          // if multiple items have the same due date, they are "stacked"
+          // if multiple items have the same due date, they are "stack"
           // which means styling making it look like there are multiple materials,
           // but only _one_ with said due date is visible.
           const loansUniqueDueDate = removeLoansWithDuplicateDueDate(
