+++ src/apps/loan-list/materials/stackable-material/stackable-material.tsx
@@ -17,6 +17,7 @@ export interface StackableMaterialProps {
   amountOfMaterialsWithDueDate?: number;
   openModal?: boolean;
   pageSize: number;
+  error: string;
 }
 
 const StackableMaterial: FC<StackableMaterialProps & MaterialProps> = ({
@@ -25,7 +26,8 @@ const StackableMaterial: FC<StackableMaterialProps & MaterialProps> = ({
   openModal,
   loan,
   stack,
-  pageSize
+  pageSize,
+  error
 }) => {
   const { open } = useModalButtonHandler();
   const [additionalMaterials] = useState(
@@ -77,6 +79,7 @@ const StackableMaterial: FC<StackableMaterialProps & MaterialProps> = ({
           additionalMaterials > 0 ? "list-reservation--stacked" : ""
         }`}
       >
+        {error && <div className="list-reservation__material">{error}</div>}
         {material && (
           <MaterialInfo
             periodical={periodical}
