+++ src/apps/reservation-list/modal/reservation-details/digital-list-details.tsx
@@ -15,7 +15,6 @@ const DigitalListDetails: FC<DigitalListDetailsProps & MaterialProps> = ({
   reservation
 }) => {
   const t = useText();
-
   const { expiryDate, dateOfReservation, pickupDeadline, state } = reservation;
 
   return (
