+++ src/core/publizon/mutator/fetcher.ts
@@ -1,7 +1,7 @@
 import { getToken, TOKEN_USER_KEY, TOKEN_LIBRARY_KEY } from "../../token";
 import { getFetcherUrl, configTypes } from "../../utils/helpers/fetcher";
 
-const defaultBaseUrl = "https://pubhub-openplatform.test.dbc.dk";
+const defaultBaseUrl = "http://publizon-mock.docker";
 
 type FetchParams =
   | string
