+++ src/core/utils/types/list-view.ts
@@ -1,1 +1,1 @@
-export type ListView = "list" | "stacked";
+export type ListView = "list" | "stack";
